Update README.md
Go to the office of the companies you have selected 
There are many companies that neither have their own office nor any transport vehicle. They also do not have their own master team to provide services like loading, unloading, rather they hire from another company.
All these things can be observed correctly only when you go to the company office. So to choose the right company, you must go to the office of those companies.
You should go to the office and talk to the spokesperson of the company on the following points.

Company Establishment : By going to the office of the packaging company selected by you, first find out when the company was established how old the company is because the older the company, the more credibility it will have. Mainly, the more the company's credibility, the more your relocation task will be successful. chances go after

Company Staff & Transportation : The strongest link of any packers and movers is the company's staff and its transportation system. You know this by going to the office. It is easy to install whether the packaging company you choose has its staff or it hires staff from a third company to complete the project. That company has its own transportation vehicle that is based on a third party. After knowing this, it becomes easier for you to choose the company of your choice.

Clear About Hidden Charges : You should make it clear from the relocation company that the quotation given by them is final that there will be other charges other than that which will be mentioned at the end of the work because often many companies charge some more money after the completion of the work even after giving the final quotation. is. So by going to the office, your doubts are cleared.

Insurance Facility: Often in work like relocation, sometimes your valuable goods get damaged. To avoid this loss, you should confirm with that company, whether any insurance facility is available them or not. Because having an insurance facility compensates for your valuable goods damaged knowingly or unintentionally.